// SDKs/twilio.com/twilio_serverless_v1/1.36.0/typescript/src/sdk/models/shared/serverlessv1servicefunctionfunctionversion.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import { FunctionVersionEnumVisibilityEnum } from "./functionversionenumvisibilityenum";


export class ServerlessV1ServiceFunctionFunctionVersion extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=function_sid" })
  functionSid?: string;

  @Metadata({ data: "json, name=links" })
  links?: Map<string, any>;

  @Metadata({ data: "json, name=path" })
  path?: string;

  @Metadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;

  @Metadata({ data: "json, name=visibility" })
  visibility?: FunctionVersionEnumVisibilityEnum;
}